Use double precision

+++ app/src/main/java/Game/GameEngine.java
@@ -31,7 +31,7 @@ public class GameEngine {
         // Generate the map
         PreciseTimer timer = new PreciseTimer();
         MapData mapData = mapGenerator.generateMap(gameSettings.getMapGenParams());
-        Log.d("*TIME generateMap:", Float.toString(timer.stop()));
+        Log.d("*TIME generateMap:", Double.toString(timer.stop()));
 
         m_gameState.territories = mapData.territories;
         m_gameState.mapData = mapData;
